Extract LocationForm component
